feat(firestore): adds useEnsureDoc hook
The purpose of this hook is to give consumer a way to make sure that a certain entity exists in DB.

Use-cases:
A user can sign in anonymously, so no registration is required as a result there is no way to call beforeUserCreated FB function.
At the same time, a new user must have a settings object attached to it.
By using this hook, we can ensure that a document for settings exists in DB and it either has default values or has already been modified by a user.